+++ src/middlewares/authMiddleware.js
@@ -1,0 +1,18 @@
+const jwt = require("jsonwebtoken");
+const config = require("../config/config");
+
+const verifyToken = (req, res, next) => {
+  const token = req.header("Authorization");
+  if (!token)
+    return res.status(401).json({ error: "Access denied. Token is required." });
+
+  try {
+    const verified = jwt.verify(token, config.jwtSecret);
+    req.user = verified;
+    next();
+  } catch (err) {
+    res.status(400).json({ error: "Invalid token." });
+  }
+};
+
+module.exports = verifyToken;
